# docs/swagger.yaml
definitions:
...
  types.APIBadRequest:
    type: object
  types.APIInternalServerError:
    type: object
...
  types.CreateSpaceReq:
    properties:
      cluster_id:
        type: string
      cover_image_url:
        type: string
      default_branch:
        example: main
        type: string
      description:
        type: string
      env:
        type: string
      labels:
        example: ""
        type: string
      license:
        example: MIT
        type: string
      name:
        example: model_name_1
        type: string
      namespace:
        example: user_or_org_name
        type: string
      nickname:
        example: model display name
        type: string
      private:
        type: boolean
      readme:
        type: string
      resource_id:
        type: integer
      sdk:
        example: "1"
        type: string
      sdk_version:
        example: v0.1
        type: string
      secrets:
        type: string
      template:
        type: string
    type: object
...
  types.Namespace:
    properties:
      avatar:
        type: string
      path:
        type: string
      type:
        description: namespace types like 'user' for normal user, and 'school', 'company'
          for orgs etc.
        type: string
    type: object
...
  types.RepoTag:
    properties:
      built_in:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      group:
        type: string
      name:
        type: string
      show_name:
        type: string
      updated_at:
        type: string
    type: object
  types.Repository:
    properties:
      http_clone_url:
        type: string
      ssh_clone_url:
        type: string
    type: object
  types.RepositorySource:
    enum:
    - opencsg
    - local
    - huggingface
    type: string
    x-enum-varnames:
    - OpenCSGSource
    - LocalSource
    - HuggingfaceSource
  types.RepositorySyncStatus:
    enum:
    - pending
    - inprogress
    - failed
    - completed
    type: string
    x-enum-varnames:
    - SyncStatusPending
    - SyncStatusInProgress
    - SyncStatusFailed
    - SyncStatusCompleted
...
  types.Response:
    properties:
      data: {}
      msg:
        type: string
    type: object
  types.ResponseWithTotal:
    properties:
      data: {}
      msg:
        type: string
      total:
        type: integer
    type: object
...
  types.Space:
    properties:
      can_manage:
        type: boolean
      can_write:
        type: boolean
      cover_image_url:
        type: string
      created_at:
        type: string
      default_branch:
        type: string
      description:
        example: ""
        type: string
      endpoint:
        description: the serving endpoint url
        example: https://localhost/spaces/myname/myspace
        type: string
      env:
        type: string
      hardware:
        type: string
      id:
        type: integer
      license:
        example: MIT
        type: string
      like_count:
        type: integer
      name:
        example: space_name_1
        type: string
      namespace:
        $ref: '#/definitions/types.Namespace'
      nickname:
        example: ""
        type: string
      path:
        example: user_or_org_name/space_name_1
        type: string
      private:
        type: boolean
      repository:
        $ref: '#/definitions/types.Repository'
      repository_id:
        type: integer
      sdk:
        description: like gradio,steamlit etc
        example: "1"
        type: string
      sdk_version:
        example: v0.1
        type: string
      secrets:
        type: string
      sku:
        type: string
      source:
        $ref: '#/definitions/types.RepositorySource'
      status:
        description: deploying, running, failed
        type: string
      svc_name:
        type: string
      sync_status:
        $ref: '#/definitions/types.RepositorySyncStatus'
      tags:
        items:
          $ref: '#/definitions/types.RepoTag'
        type: array
      template:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/types.User'
      user_likes:
        type: boolean
      username:
        example: creator_user_name
        type: string
    type: object
...
  types.UpdateSpaceReq:
    properties:
      cover_image_url:
        type: string
      description:
        type: string
      env:
        type: string
      nickname:
        example: model display name
        type: string
      private:
        example: false
        type: boolean
      resource_id:
        type: integer
      sdk:
        example: "1"
        type: string
      sdk_version:
        example: v0.1
        type: string
      secrets:
        type: string
      template:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /spaces:
    get:
      consumes:
      - application/json
      description: get spaces visible to current user
      parameters:
      - description: current user
        in: query
        name: current_user
        type: string
      - description: search text
        in: query
        name: search
        type: string
      - description: filter by task tag
        in: query
        name: task_tag
        type: string
      - description: filter by framework tag
        in: query
        name: framework_tag
        type: string
      - description: filter by license tag
        in: query
        name: license_tag
        type: string
      - description: filter by language tag
        in: query
        name: language_tag
        type: string
      - description: sort by
        in: query
        name: sort
        type: string
      - description: source
        enum:
        - opencsg
        - huggingface
        - local
        in: query
        name: source
        type: string
      - default: 20
        description: per
        in: query
        name: per
        type: integer
      - default: 1
        description: per page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.ResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.Space'
                  type: array
                total:
                  type: integer
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Get spaces visible to current user
      tags:
      - Space
    post:
      consumes:
      - application/json
      description: create a new space
      parameters:
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.CreateSpaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Space'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Create a new space
      tags:
      - Space
  /spaces/{namespace}/{name}:
    delete:
      consumes:
      - application/json
      description: delete a exists space
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Delete a exists space
      tags:
      - Space
    get:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Space'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: show space detail
      tags:
      - Space
    put:
      consumes:
      - application/json
      description: update a exists space
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/types.UpdateSpaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.Space'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - ApiKey: []
      summary: Update a exists space
      tags:
      - Space
  /spaces/{namespace}/{name}/logs:
    get:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: get space logs
      tags:
      - Space
  /spaces/{namespace}/{name}/run:
    post:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: run space app
      tags:
      - Space
  /spaces/{namespace}/{name}/status:
    get:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: get space status
      tags:
      - Space
  /spaces/{namespace}/{name}/stop:
    post:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: current_user
        in: query
        name: current_user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: stop space app
      tags:
      - Space
  /spaces/{namespace}/{name}/wakeup:
    post:
      consumes:
      - application/json
      parameters:
      - description: namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/types.APIBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.APIInternalServerError'
      security:
      - JWT token: []
      summary: wake up space app
      tags:
      - Space
...
securityDefinitions:
  ApiKey:
    description: Bearer token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"